requirements.txt
click>=8.0.3
tqdm>=4.62.3
httplib2>=0.20.1
google-auth-httplib2
google-auth-oauthlib
google-api-python-client
pycryptodomex